Done updating and restructuring to 1.3 NOT FULLY TESTED
