+++ studious-octo-spork/.meteor/release
@@ -1,1 +1,1 @@
-METEOR@1.3
+METEOR@1.3.1
