+++ studious-octo-spork/.meteor/versions
@@ -1,96 +1,96 @@
-accounts-base@1.2.4
-accounts-password@1.1.6
-accounts-ui@1.1.7
-accounts-ui-unstyled@1.1.10
-allow-deny@1.0.2
-autopublish@1.0.5
-autoupdate@1.2.6
-babel-compiler@6.5.2
-babel-runtime@0.1.6
-base64@1.0.6
-binary-heap@1.0.6
-blaze@2.1.5
-blaze-html-templates@1.0.2
-blaze-tools@1.0.6
-boilerplate-generator@1.0.6
-caching-compiler@1.0.2
-caching-html-compiler@1.0.4
-callback-hook@1.0.6
-check@1.1.2
-coffeescript@1.0.15
+accounts-base@1.2.5
+accounts-password@1.1.7
+accounts-ui@1.1.8
+accounts-ui-unstyled@1.1.11
+allow-deny@1.0.3
+autopublish@1.0.6
+autoupdate@1.2.7
+babel-compiler@6.6.1
+babel-runtime@0.1.7
+base64@1.0.7
+binary-heap@1.0.7
+blaze@2.1.6
+blaze-html-templates@1.0.3
+blaze-tools@1.0.7
+boilerplate-generator@1.0.7
+caching-compiler@1.0.3
+caching-html-compiler@1.0.5
+callback-hook@1.0.7
+check@1.1.3
+coffeescript@1.0.16
 dandv:caret-position@2.1.1
-ddp@1.2.3
-ddp-client@1.2.3
-ddp-common@1.2.3
-ddp-rate-limiter@1.0.2
-ddp-server@1.2.4
-deps@1.0.10
-diff-sequence@1.0.3
-ecmascript@0.4.1
-ecmascript-runtime@0.2.8
-ejson@1.0.9
-email@1.0.10
-es5-shim@4.5.8
-fastclick@1.0.9
-geojson-utils@1.0.6
-hot-code-push@1.0.2
-html-tools@1.0.7
-htmljs@1.0.7
-http@1.1.3
-id-map@1.0.5
-insecure@1.0.5
-jquery@1.11.6
+ddp@1.2.4
+ddp-client@1.2.4
+ddp-common@1.2.4
+ddp-rate-limiter@1.0.3
+ddp-server@1.2.5
+deps@1.0.11
+diff-sequence@1.0.4
+ecmascript@0.4.2
+ecmascript-runtime@0.2.9
+ejson@1.0.10
+email@1.0.11
+es5-shim@4.5.9
+fastclick@1.0.10
+geojson-utils@1.0.7
+hot-code-push@1.0.3
+html-tools@1.0.8
+htmljs@1.0.8
+http@1.1.4
+id-map@1.0.6
+insecure@1.0.6
+jquery@1.11.7
 kadira:blaze-layout@2.3.0
 kadira:flow-router@2.11.0
-launch-screen@1.0.8
-less@2.5.6
-livedata@1.0.16
-localstorage@1.0.7
-logging@1.0.10
-meteor@1.1.12
-meteor-base@1.0.2
-minifier-css@1.1.9
-minifier-js@1.1.9
-minimongo@1.0.12
+launch-screen@1.0.10
+less@2.5.7
+livedata@1.0.17
+localstorage@1.0.8
+logging@1.0.11
+meteor@1.1.13
+meteor-base@1.0.3
+minifier-css@1.1.10
+minifier-js@1.1.10
+minimongo@1.0.13
 mizzao:autocomplete@0.5.1
-mobile-experience@1.0.2
-mobile-status-bar@1.0.10
-modules@0.5.1
-modules-runtime@0.6.1
-mongo@1.1.5
-mongo-id@1.0.2
+mobile-experience@1.0.3
+mobile-status-bar@1.0.11
+modules@0.5.2
+modules-runtime@0.6.2
+mongo@1.1.6
+mongo-id@1.0.3
 npm-bcrypt@0.7.8_2
-npm-mongo@1.4.41
-observe-sequence@1.0.9
-ordered-dict@1.0.5
+npm-mongo@1.4.42
+observe-sequence@1.0.10
+ordered-dict@1.0.6
 practicalmeteor:chai@2.1.0_1
 practicalmeteor:loglevel@1.2.0_2
 practicalmeteor:mocha@2.4.5_1
 practicalmeteor:mocha-core@0.1.4
 practicalmeteor:sinon@1.14.1_2
-promise@0.6.5
-random@1.0.7
-rate-limit@1.0.2
-reactive-dict@1.1.5
-reactive-var@1.0.7
-reload@1.1.6
-retry@1.0.5
-routepolicy@1.0.8
-service-configuration@1.0.7
-session@1.1.3
-sha@1.0.5
-spacebars@1.0.9
-spacebars-compiler@1.0.9
-srp@1.0.6
-standard-minifier-css@1.0.4
-standard-minifier-js@1.0.4
-templating@1.1.7
-templating-tools@1.0.2
+promise@0.6.6
+random@1.0.8
+rate-limit@1.0.3
+reactive-dict@1.1.6
+reactive-var@1.0.8
+reload@1.1.7
+retry@1.0.6
+routepolicy@1.0.9
+service-configuration@1.0.8
+session@1.1.4
+sha@1.0.6
+spacebars@1.0.10
+spacebars-compiler@1.0.10
+srp@1.0.7
+standard-minifier-css@1.0.5
+standard-minifier-js@1.0.5
+templating@1.1.8
+templating-tools@1.0.3
 tmeasday:test-reporter-helpers@0.2.1
-tracker@1.0.11
+tracker@1.0.12
 twbs:bootstrap@3.3.6
-ui@1.0.9
-underscore@1.0.6
-url@1.0.7
-webapp@1.2.6
-webapp-hashing@1.0.7
+ui@1.0.10
+underscore@1.0.7
+url@1.0.8
+webapp@1.2.7
+webapp-hashing@1.0.8
