+++ studious-octo-spork/client/main.js
@@ -1,0 +1,2 @@
+import '../imports/startup/client';
+import '../imports/ui/body.js';
