+++ studious-octo-spork/imports/api/messaging.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo';
+
+export new Mongo.Collection('messages');
