+++ studious-octo-spork/imports/startup/client/index.js
@@ -1,0 +1,2 @@
+import './accounts-config.js';
+import './routes.js';
