+++ studious-octo-spork/imports/ui/contact.html
@@ -1,0 +1,54 @@
+<template name="contact">
+  <div class="jumbotron">
+    <form role="form" id="cntct" class="container-fluid testemail">
+      <h3 class="text-center">Contact us and we'll get back to you within 24 hours.</h3>
+      <br>
+      <div class="row">
+        <div class="col-md-4">
+          <p><span class="glyphicon glyphicon-map-marker"></span>Athens, GR</p>
+          <p><span class="glyphicon glyphicon-phone"></span>Phone: +30 6912345678</p>
+          <p><span class="glyphicon glyphicon-envelope"></span>Email: mail@mail.com</p>
+        </div>
+        <div class="col-md-8">
+          <div class="row">
+            <div class="col-sm-6 form-group">
+              <input class="form-control" id="name" name="name" placeholder="Name" type="text" required>
+            </div>
+            <div class="col-sm-6 form-group">
+              <input class="form-control" id="email" name="email" placeholder="Email" type="text" required>
+            </div>
+            <textarea class="form-control" id="comments" name="comments" placeholder="Comment" rows="5"></textarea><br>
+            <div class="row">
+              <div class="col-sm-12 form-group">
+                <button class="btn btn-default pull-right" type="submit">Send</button>
+              </div>
+            </div>
+          </div>
+          <p class='thnx'></p>
+        </div>
+      </div>
+    </form>
+  </div>
+  <script type='text/javascript'>
+    $(document).ready( function() {
+      $( 'form' ).submit( function() {
+        var noempty = true;
+        if ( this.name.value == null || this.name.value == "" ) {
+          noempty = false;
+        }
+        if ( this.email.value == null || this.email.value == "" ) {
+          noempty = false;
+        }
+        if ( this.comments.value == null || this.comments.value == "" ) {
+          noempty = false;
+        }
+        if ( !noempty ) {
+          alert("There is empty input");
+          return noempty;
+        }
+        $( '.thnx' ).html( '<strong>Your comment has been received, Thank you!</strong>' ).show().fadeOut( 10000 );
+      });
+    });
+  </script>
+</template>
+
