+++ studious-octo-spork/imports/ui/contact.js
@@ -1,0 +1,18 @@
+import { Meteor } from 'meteor/meteor';
+import { Template } from 'meteor/templating';
+
+import './contact.html';
+
+Template.contact.events({
+  'submit .testemail': function (event) {
+    event.preventDefault();
+    Meteor.call('sendTestEmail',{
+      name: event.target.name.value,
+      email: event.target.email.value,
+      comment: event.target.comments.value
+    });
+    event.target.name.value = '';
+    event.target.email.value = '';
+    event.target.comments.value = '';
+  }
+});
