+++ studious-octo-spork/imports/ui/dashboard.html
@@ -1,0 +1,65 @@
+<template name="dashboard">
+  <div class="jumbotron">
+    <div class="container-fluid">
+      {{#if formData}}
+        {{#with formData}}
+          <form role="form" class="form-horizontal update-profile" method="post">
+            <div class="form-group">
+              <label for="username" class="control-label col-sm-1">Username: </label>
+              <div class="col-sm-2">
+                <input type="text" class="form-control" id="username" name="username" value={{username}}>
+              </div>
+            </div>
+            <div class="form-group">
+              <label for="occupation" class="control-label col-sm-1">Occupation: </label>
+              <div class="col-sm-2">
+                {{>inputAutocomplete settings=settings id='occupation' class='form-control' name='occupation' value=occupation }}
+              </div>
+            </div>
+            <div class="form-group">
+              <label for="description" class="control-label col-sm-1">Description: </label>
+              <div class="col-sm-2">
+                <input type="text" class="form-control" id="description" name="description" value={{description}}>
+              </div>
+            </div>
+            <div class="form-group">
+              <div class="col-sm-2 col-sm-8">
+                <button type="submit" class="btn btn-default">Save</button>
+              </div>
+            </div>
+          </form>
+        {{/with}}
+      {{else}}
+        <p>You are not registered yet.</p>
+      {{/if}}
+    </div>
+  </div>
+  {{#if currentUser}}
+    <div class="container-fluid">
+      <h3>Messages</h3>
+      <ul>
+        {{#each userMessages}}
+          <li>
+            {{from}}: {{message}}
+            <button class="toggle-read">
+              {{#if read}}
+                <span class="glyphicon glyphicon-eye-open"></span>
+              {{else}}
+                <span class="glyphicon glyphicon-eye-close"></span>
+              {{/if}}
+            </button>
+            <button>
+              <a href="/users/{{from}}/contact">
+                <span class="glyphicon glyphicon-pencil"></span>
+              </a>
+            </button>
+            <button class="delete">
+              <span class="glyphicon glyphicon-remove"></span>
+            </button>
+          </li>
+        {{/each}}
+      </ul>
+    </div>
+  {{/if}}
+</template>
+
