+++ studious-octo-spork/imports/ui/users.html
@@ -1,0 +1,122 @@
+<template name="usersIndex">
+  <div class="jumbotron">
+    <div class="container-fluid">
+      <h2>Get your job done.<br />In a fair manner.</h2>
+    </div>
+  </div>
+  <div class="container-fluid">
+    <p>An overview of all the users.</p>
+	  {{# if users}}
+		  <ul>
+		  {{#each users}}
+        <a href="{{path(username)}}">
+          <li>{{username}}: {{profile.occupation}}</li>
+        </a>
+		  {{/each}}
+		  </ul>
+	  {{/if}}
+  </div>
+  <div class='container'>
+    <div class='col-md-6'>
+      <form class='findOccupation'>
+        <input class="col-md-6" type='text' name='work' placeholder='Are you looking for an occupation?' />
+      </form>
+    </div>
+    <div class='col-md-6'>
+      <form class='findDescription'>
+        <input class="col-md-6" type='text' name='description' placeholder='Give me a description!' />
+      </form>
+    </div>
+  </div>
+  <div class='container'>
+    <div class='col-md-6'>
+	    {{#if usersSearchWithOccupation}}
+	      <ul>
+		      {{#each usersSearchWithOccupation}}
+            <a href="{{path(username)}}">
+              <li>{{username}}: {{profile.occupation}}</li>
+            </a>
+          {{/each}}
+		    </ul>
+      {{/if}}
+    </div>
+    <div class='col-md-6'>
+	    {{#if usersSearchWithDescription}}
+		    <ul>
+	        {{#each usersSearchWithDescription}}
+            <a href="{{path(username)}}">
+              <li>{{username}}: {{profile.occupation}}</li>
+            </a>
+		      {{/each}}
+		    </ul>
+      {{/if}}
+    </div>
+  </div>
+</template>
+
+<template name="usersByUsername">
+  <div class="jumbotron">
+    <div class="container-fluid">
+      {{#if userData}}
+        {{#with userData}}
+          <h3>user {{username}}</h3>
+          <p>
+            {{#if occupation}}
+              Occupation: {{occupation}}<br />
+            {{else}}
+              User has not registered an occupation yet.<br />
+            {{/if}}
+            {{#if description}}
+              {{description}}<br />
+            {{/if}}
+            <a href="/users/{{targetUsername}}/contact">contact user</a>
+          </p>
+        {{/with}}
+      {{else}}
+        <p>User {{targetUsername}} does not exist.</p>
+      {{/if}}
+    </div>
+  </div>
+</template>
+
+<template name="usersContactByUsername">
+  <div class="jumbotron">
+    <div class="container-fluid">
+      {{#if currentUser}}
+        <form role="form" class="form-horizontal send-message" method="post">
+          <div class="form-group">
+            <label for="message" class="control-label col-sm-3">
+              Your Message for {{targetUsername}}:
+            </label>
+            <div class="col-sm-4">
+              <input type="text" class="form-control" id="message" name="message" value="Hello {{targetUsername}}, ">
+            </div>
+          </div>
+          <div class="form-group">
+            <div class="col-sm-offset-3 col-sm-4">
+              <button type="submit" class="btn btn-default">Send</button>
+            </div>
+          </div>
+        </form>
+        <p class='received text-center'></p>
+      {{else}}
+        <h2>Sorry, you need to login to access this page</h2>
+      {{/if}}
+    </div>
+  </div>
+  <script type='text/javascript'>
+    $(document).ready( function() {
+      $( 'form' ).submit( function() {
+        var noempty = true;
+        if ( this.message.value == null || this.message.value == "" ) {
+          noempty = false;
+        }
+        if ( !noempty ) {
+          alert("Your message is empty");
+          return noempty;
+        }
+        $( '.received' ).html( '<strong>Your message has been received</strong>' ).show().fadeOut( 10000 );
+      });
+    });
+  </script>
+</template>
