+++ studious-octo-spork/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "studious-octo-spork",
+  "private": true,
+  "scripts": {
+    "start": "meteor run"
+  },
+  "dependencies": {
+    "meteor-node-stubs": "~0.2.0"
+  }
+}
